pyproject.toml
[project]
name = "demo-django-api"
version = "0.1.0"
description = "Django API Demo Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<6.0",
    "celery>=5.3.0",
    "django-celery-beat>=2.5.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "django-split-settings>=1.1.0",
    "gunicorn>=21.2.0",
    "django-debug-toolbar>=3.7.0",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.4",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DJ",  # flake8-django
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["config", "request"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*.py" = ["E501", "DJ01"]
"*/tests/*.py" = ["B008"]
"gunicorn.conf.py" = ["N999"]
